# Implement single event listener for Touchable components to boost performance