## Summary

Each Touchable component needlessly creates own event listener, which, after focus transition, results in multiple handlers called just to check if event should be even handled by particular component. Performance of focus transition can be improved.

## Changelog

[General] [Changed] - Improved focus transition performance of Touchable components by introducing single Event Listener for all Touchables, instead of one Event Listener per Touchable.

## Test Plan

I've prepared simple TV app with improvement implemented as a patch to check and measure performance improvement. Rough measurement of focus transition between tiles on AndroidTV emulator indicates that time between onBlur -> onFocus pair reduced from ~11 ms to ~7 ms on average, and peak value form ~48 ms to ~14 ms.
https://github.com/Zahoq/TouchableFocusImprovementTestApp