@douglowder 2022-01-24
This looks great. Could you also modify `Libraries/Components/Pressable/Pressable.js` to use the new focus handler?