@Zahoq 2022-01-24
> This looks great. Could you also modify `Libraries/Components/Pressable/Pressable.js` to use the new focus handler?
Sure. I'll move initialization of `TVFocusEventHandler` instance to `TVFocusEventHandler.js` so both Touchables and Pressables use same instance of that class.